+++ ClientApp/src/app/app-routing.module.ts
@@ -6,7 +6,8 @@ import { HomeComponent } from './home/home.component';
 const routes: Routes = [
     { path: '', redirectTo: 'home', pathMatch: 'full' },
     { path: 'home', component: HomeComponent },
-    { path: 'resource', loadChildren: './resource/resource.module#ResourceModule' }
+    { path: 'resource', loadChildren: './resource/resource.module#ResourceModule' },
+    { path: '**', redirectTo: 'home' },
 ];
 
 @NgModule({