+++ src/environments/environment.hmr.ts
@@ -1,4 +1,0 @@
-export const environment: any = {
-    production: false,
-    hmr: true
-};