+++ src/environments/environment.prod.ts
@@ -1,4 +1,0 @@
-export const environment: any = {
-    production: true,
-    hmr: false
-};