+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 ## Angular CLI ASP.NET Core Changelog
 
+<a name="Oct 22, 2018"></a>
+### Oct 22, 2018
+* Upgrade to Angular v7, .NET Core 2.1 and switch to SpaServices
+
 <a name="May 7, 2018"></a>
 ### May 7, 2018
 * Upgrade to Angular v6